refactor: optimize vector index building and querying performance
