+++ hugegraph-llm/src/hugegraph_llm/operators/index_op/build_semantic_index.py
@@ -20,13 +20,12 @@ import asyncio
 import os
 from typing import Any, Dict
 
-from tqdm import tqdm
-
 from hugegraph_llm.config import resource_path, huge_settings
 from hugegraph_llm.indices.vector_index import VectorIndex
 from hugegraph_llm.models.embeddings.base import BaseEmbedding
 from hugegraph_llm.operators.hugegraph_op.schema_manager import SchemaManager
 from hugegraph_llm.utils.log import log
+from hugegraph_llm.utils.embedding_utils import get_embeddings_parallel
 
 
 class BuildSemanticIndex:
@@ -38,34 +37,6 @@ class BuildSemanticIndex:
 
     def _extract_names(self, vertices: list[str]) -> list[str]:
         return [v.split(":")[1] for v in vertices]
-
-    async def _get_embeddings_parallel(self, vids: list[str]) -> list[Any]:
-        sem = asyncio.Semaphore(10)
-        batch_size = 1000
-
-        # TODO: refactor the logic here (call async method)
-        async def get_embeddings_with_semaphore(vid_list: list[str]) -> Any:
-            # Executes sync embedding method in a thread pool via loop.run_in_executor, combining async programming
-            # with multi-threading capabilities.
-            # This pattern avoids blocking the event loop and prepares for a future fully async pipeline.
-            async with sem:
-                loop = asyncio.get_running_loop()
-                # FIXME: [PR-238] add & use async_get_texts_embedding instead of sync method
-                return await loop.run_in_executor(None, self.embedding.get_texts_embeddings, vid_list)
-
-        # Split vids into batches of size batch_size
-        vid_batches = [vids[i:i + batch_size] for i in range(0, len(vids), batch_size)]
-
-        # Create tasks for each batch
-        tasks = [get_embeddings_with_semaphore(batch) for batch in vid_batches]
-
-        embeddings = []
-        with tqdm(total=len(tasks)) as pbar:
-            for future in asyncio.as_completed(tasks):
-                batch_embeddings = await future
-                embeddings.extend(batch_embeddings)  # Extend the list with batch results
-                pbar.update(1)
-        return embeddings
 
     def run(self, context: Dict[str, Any]) -> Dict[str, Any]:
         vertexlabels = self.sm.schema.getSchema()["vertexlabels"]
@@ -80,7 +51,7 @@ class BuildSemanticIndex:
 
         if added_vids:
             vids_to_process = self._extract_names(added_vids) if all_pk_flag else added_vids
-            added_embeddings = asyncio.run(self._get_embeddings_parallel(vids_to_process))
+            added_embeddings = asyncio.run(get_embeddings_parallel(self.embedding, vids_to_process))
             log.info("Building vector index for %s vertices...", len(added_vids))
             self.vid_index.add(added_embeddings, added_vids)
             self.vid_index.to_index_file(self.index_dir)
