+++ hugegraph-llm/src/hugegraph_llm/operators/index_op/vector_index_query.py
@@ -34,7 +34,7 @@ class VectorIndexQuery:
 
     def run(self, context: Dict[str, Any]) -> Dict[str, Any]:
         query = context.get("query")
-        query_embedding = self.embedding.get_text_embedding(query)
+        query_embedding = self.embedding.get_texts_embeddings([query])[0]
         # TODO: why set dis_threshold=2?
         results = self.vector_index.search(query_embedding, self.topk, dis_threshold=2)
         # TODO: check format results
